python-boilerplate/requirements.txt
coverage==5.5
logzero==1.7.0
pycodestyle==2.7.0
pyflakes==2.3.0
pylint==2.7.2
flake8==3.9.0
python-coveralls==2.9.3

# Data and trading packages
databento==0.64.0
alpaca-py==0.43.2
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.0.2
scipy==1.13.1
matplotlib==3.9.4
duckdb==1.1.3
openpyxl==3.1.5
requests==2.31.0
yfinance==0.2.66

# Jupyter packages
jupyter==1.1.1
ipykernel==6.31.0
